pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "betanumerics"
version = "0.1.0"
description = "A web application for generating unique betanumeric identifiers"
authors = [{name = "Betanumerics Team"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Flask>=3.0.0",
    "Werkzeug>=3.0.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 